pyproject.toml
[project]
name = "betmock"
version = "0.1.0"
description = ""
authors = [
    {name = "Javier Iglesias",email = "javo.iglesias@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "rich (>=14.0.0,<15.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
